+++ packages/loaders/src/SvgLoader/SvgLoader.styled.tsx
@@ -1,8 +1,8 @@
-import { defaultTheme } from '@medly-components/theme';
+import { defaultTheme, LoaderSizes } from '@medly-components/theme';
 import { InjectClassName, styled } from '@medly-components/utils';
 import { SvgLoaderProps } from './types';
 
-export const SvgLoaderStyled = styled(InjectClassName)<SvgLoaderProps>`
+export const SvgLoaderStyled = styled(InjectClassName)<SvgLoaderProps & { size: LoaderSizes }>`
     font-size: ${({ theme, size }) => theme.loader.sizes[size]};
     * {
         fill: ${({ theme, color }) => color || theme.loader.defaultColor};
@@ -10,7 +10,6 @@ export const SvgLoaderStyled = styled(InjectClassName)<SvgLoaderProps>`
 `;
 
 SvgLoaderStyled.displayName = 'SvgLoader';
-
 SvgLoaderStyled.defaultProps = {
     size: 'S',
     theme: defaultTheme