+++ packages/theme/src/core/colors/medlyGreen.ts
@@ -4,7 +4,10 @@ const medlyGreen: Color = {
     200: '#D8F8F2',
     300: '#9EEDE0',
     400: '#59E0C9',
-    500: '#3CDBC0'
+    500: '#3CDBC0',
+    600: '#3CDBC0',
+    700: '#3CDBC0',
+    800: '#3CDBC0'
 };
 
 export default medlyGreen;