+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.30.8](https://github.com/medly/medly-components/compare/@medly-components/forms@1.30.7...@medly-components/forms@1.30.8) (2021-09-21)
+
+**Note:** Version bump only for package @medly-components/forms
+
+
+
+
+
+## [1.30.7](https://github.com/medly/medly-components/compare/@medly-components/forms@1.30.6...@medly-components/forms@1.30.7) (2021-09-21)
+
+**Note:** Version bump only for package @medly-components/forms
+
+
+
+
+
+## [1.30.6](https://github.com/medly/medly-components/compare/@medly-components/forms@1.30.5...@medly-components/forms@1.30.6) (2021-09-21)
+
+**Note:** Version bump only for package @medly-components/forms
+
+
+
+
+
 ## [1.30.5](https://github.com/medly/medly-components/compare/@medly-components/forms@1.30.4...@medly-components/forms@1.30.5) (2021-09-20)
 
 **Note:** Version bump only for package @medly-components/forms