+++ packages/loaders/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.35](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.2.34...@medly-components/loaders@1.2.35) (2021-09-21)
+
+**Note:** Version bump only for package @medly-components/loaders
+
+
+
+
+
+## [1.2.34](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.2.33...@medly-components/loaders@1.2.34) (2021-09-21)
+
+**Note:** Version bump only for package @medly-components/loaders
+
+
+
+
+
 ## [1.2.33](https://github.com/medly/medly-components/compare/@medly-components/loaders@1.2.32...@medly-components/loaders@1.2.33) (2021-09-19)
 
 **Note:** Version bump only for package @medly-components/loaders